+++ crates/account-decoder/Cargo.toml
@@ -17,4 +17,4 @@ solders-macros = { workspace = true }
 zstd = "0.11.2"
 base64 = { workspace = true }
 bs58 = { workspace = true }
-solders-traits = { workspace = true }
+solders-traits-core = { workspace = true }
