+++ crates/account/Cargo.toml
@@ -13,6 +13,6 @@ derive_more = { workspace = true }
 solana-sdk = { workspace = true }
 solders-macros = { workspace = true }
 solders-pubkey = { workspace = true }
-solders-traits = { workspace = true }
+solders-traits-core = { workspace = true }
 solders-account-decoder = { workspace = true }
 base64 = { workspace = true }
