+++ crates/account/src/lib.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 use solana_sdk::{account::Account as AccountOriginal, clock::Epoch};
 use solders_macros::{common_methods, richcmp_eq_only};
 use solders_pubkey::Pubkey;
-use solders_traits::{
+use solders_traits_core::{
     impl_display, py_from_bytes_general_via_bincode, pybytes_general_via_bincode,
     RichcmpEqualityOnly,
 };
@@ -99,7 +99,7 @@ impl_display!(Account);
 pybytes_general_via_bincode!(Account);
 py_from_bytes_general_via_bincode!(Account);
 
-solders_traits::common_methods_default!(Account);
+solders_traits_core::common_methods_default!(Account);
 impl RichcmpEqualityOnly for Account {}
 
 impl TryFrom<UiAccount> for Account {
@@ -186,7 +186,7 @@ impl std::fmt::Display for AccountJSON {
 pybytes_general_via_bincode!(AccountJSON);
 py_from_bytes_general_via_bincode!(AccountJSON);
 
-solders_traits::common_methods_default!(AccountJSON);
+solders_traits_core::common_methods_default!(AccountJSON);
 impl RichcmpEqualityOnly for AccountJSON {}
 
 impl TryFrom<UiAccount> for AccountJSON {
