+++ crates/keypair/src/lib.rs
@@ -11,11 +11,11 @@ use solders_macros::{common_methods, pyhash, richcmp_signer};
 use solders_pubkey::Pubkey;
 use solders_signature::Signature;
 
-use solders_traits::{
-    handle_py_value_err, impl_display, impl_signer_hash, CommonMethods, CommonMethodsCore,
-    PyBytesGeneral, PyFromBytesGeneral, PyHash, RichcmpSigner, SignerTraitWrapper,
-    ToSignerOriginal,
+use solders_traits_core::{
+    impl_display, CommonMethods, CommonMethodsCore,
+    PyBytesGeneral, PyFromBytesGeneral, PyHash, handle_py_value_err
 };
+use solders_traits::{ToSignerOriginal, SignerTraitWrapper, RichcmpSigner, impl_signer_hash};
 
 pub mod null_signer;
 pub mod presigner;
