+++ crates/keypair/src/null_signer.rs
@@ -6,10 +6,11 @@ use solders_macros::{common_methods, pyhash, richcmp_signer};
 use solders_pubkey::Pubkey;
 use solders_signature::Signature;
 
-use solders_traits::{
-    impl_display, impl_signer_hash, CommonMethodsCore, PyBytesGeneral, PyFromBytesGeneral, PyHash,
-    RichcmpSigner, SignerTraitWrapper, ToSignerOriginal,
+use solders_traits_core::{
+    impl_display, CommonMethodsCore,
+    PyBytesGeneral, PyFromBytesGeneral, PyHash
 };
+use solders_traits::{ToSignerOriginal, SignerTraitWrapper, RichcmpSigner, impl_signer_hash};
 
 mod null_signer_serde {
     use serde::{self, Deserialize, Deserializer, Serializer};
@@ -116,7 +117,7 @@ impl PyFromBytesGeneral for NullSigner {
     }
 }
 
-solders_traits::common_methods_default!(NullSigner);
+solders_traits_core::common_methods_default!(NullSigner);
 
 impl ToSignerOriginal for NullSigner {
     fn to_inner(&self) -> Box<dyn SignerTrait> {
