+++ crates/message/Cargo.toml
@@ -9,6 +9,7 @@ repository = "https://github.com/kevinheavey/solders"
 
 [dependencies]
 pyo3 = { workspace = true, features = ["macros"] }
+solders-traits-core = { workspace = true }
 solders-traits = { workspace = true }
 solders-macros = { workspace = true }
 solana-program = { workspace = true }
