+++ crates/rpc-common/Cargo.toml
@@ -11,7 +11,7 @@ pyo3 = { workspace = true }
 serde = { workspace = true }
 derive_more = { workspace = true }
 solders-macros = { workspace = true }
-solders-traits = { workspace = true }
+solders-traits-core = { workspace = true }
 solders-account-decoder = { workspace = true }
 solders-account = { workspace = true }
 solders-transaction-status = { workspace = true }
