+++ crates/rpc-common/src/lib.rs
@@ -14,15 +14,15 @@ use solders_account::Account;
 #[macro_export]
 macro_rules! response_data_boilerplate {
     ($name:ident) => {
-        impl solders_traits::RichcmpEqualityOnly for $name {}
+        impl solders_traits_core::RichcmpEqualityOnly for $name {}
         impl std::fmt::Display for $name {
             fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
                 write!(f, "{:?}", self)
             }
         }
-        solders_traits::pybytes_general_via_bincode!($name);
-        solders_traits::py_from_bytes_general_via_bincode!($name);
-        solders_traits::common_methods_default!($name);
+        solders_traits_core::pybytes_general_via_bincode!($name);
+        solders_traits_core::py_from_bytes_general_via_bincode!($name);
+        solders_traits_core::common_methods_default!($name);
     };
 }
