+++ crates/rpc-config/Cargo.toml
@@ -10,7 +10,7 @@ repository = "https://github.com/kevinheavey/solders"
 pyo3 = { workspace = true }
 serde = { workspace = true }
 solders-macros = { workspace = true }
-solders-traits = { workspace = true }
+solders-traits-core = { workspace = true }
 solders-account-decoder = { workspace = true }
 solders-pubkey = { workspace = true }
 solders-hash = { workspace = true }
