+++ crates/rpc-filter/Cargo.toml
@@ -11,7 +11,7 @@ pyo3 = { workspace = true }
 serde = { workspace = true }
 derive_more = { workspace = true }
 solders-macros = { workspace = true }
-solders-traits = { workspace = true }
+solders-traits-core = { workspace = true }
 thiserror = { workspace = true }
 bs58 = { workspace = true }
 base64 = { workspace = true }
