+++ crates/rpc-filter/src/lib.rs
@@ -9,7 +9,7 @@ use tmp_filter::{
 use derive_more::{From, Into};
 use solders_macros::{common_methods, enum_original_mapping, richcmp_eq_only, EnumIntoPy};
 
-use solders_traits::{
+use solders_traits_core::{
     impl_display, py_from_bytes_general_via_bincode, pybytes_general_via_bincode,
     RichcmpEqualityOnly,
 };
@@ -97,7 +97,7 @@ impl Memcmp {
 }
 
 impl RichcmpEqualityOnly for Memcmp {}
-solders_traits::common_methods_default!(Memcmp);
+solders_traits_core::common_methods_default!(Memcmp);
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, FromPyObject, EnumIntoPy)]
 #[serde(rename_all = "camelCase")]
