+++ crates/rpc-requests/Cargo.toml
@@ -15,6 +15,7 @@ serde_json = { workspace = true }
 derive_more = { workspace = true }
 solders-hash = { workspace = true }
 solders-macros = { workspace = true }
+solders-traits-core = { workspace = true }
 solders-traits = { workspace = true }
 solders-transaction-status = { workspace = true }
 solders-commitment-config = { workspace = true }
