+++ crates/transaction-status/Cargo.toml
@@ -18,7 +18,7 @@ solders-signature = { workspace = true }
 solders-macros = { workspace = true }
 base64 = { workspace = true }
 bs58 = { workspace = true }
-solders-traits = { workspace = true }
+solders-traits-core = { workspace = true }
 solders-account-decoder = { workspace = true }
 solders-hash = { workspace = true }
 solders-message = { workspace = true }
