+++ crates/transaction/src/lib.rs
@@ -14,10 +14,11 @@ use solana_sdk::{
 };
 use solders_macros::{common_methods, richcmp_eq_only, EnumIntoPy};
 use solders_pubkey::{convert_optional_pubkey, Pubkey};
-use solders_traits::{
-    handle_py_err, impl_display, py_from_bytes_general_via_bincode, pybytes_general_via_bincode,
+use solders_traits_core::{
+    impl_display, py_from_bytes_general_via_bincode, pybytes_general_via_bincode,
     CommonMethodsCore, RichcmpEqualityOnly,
 };
+use solders_traits::handle_py_err;
 
 use solders_hash::Hash as SolderHash;
 use solders_instruction::{convert_instructions, CompiledInstruction, Instruction};
@@ -47,7 +48,7 @@ impl RichcmpEqualityOnly for VersionedTransaction {}
 pybytes_general_via_bincode!(VersionedTransaction);
 py_from_bytes_general_via_bincode!(VersionedTransaction);
 impl_display!(VersionedTransaction);
-solders_traits::common_methods_default!(VersionedTransaction);
+solders_traits_core::common_methods_default!(VersionedTransaction);
 
 #[richcmp_eq_only]
 #[common_methods]
@@ -694,7 +695,7 @@ impl RichcmpEqualityOnly for Transaction {}
 pybytes_general_via_bincode!(Transaction);
 py_from_bytes_general_via_bincode!(Transaction);
 impl_display!(Transaction);
-solders_traits::common_methods_default!(Transaction);
+solders_traits_core::common_methods_default!(Transaction);
 
 impl AsRef<TransactionOriginal> for Transaction {
     fn as_ref(&self) -> &TransactionOriginal {
@@ -721,7 +722,7 @@ impl Legacy {
     ) -> pyo3::prelude::PyResult<bool> {
         // we override the default impl since it implicitly casts to in which causes problems when transaction
         // version is represented as `Legacy | int`.
-        solders_traits::RichcmpEqualityOnly::richcmp(self, other, op)
+        solders_traits_core::RichcmpEqualityOnly::richcmp(self, other, op)
     }
 }
